optimized layouts file to not have direct import to wagmi
